feat: checkupList 목록데이터에 totalCount추가, CheckupList의 id타입 Long으로 변경

+++ sebakwi-be/src/main/java/com/ssafy/sebakwi/checkupList/controller/CheckupListController.java
@@ -49,7 +49,7 @@ public class CheckupListController {
 
 
     @GetMapping("/{checkupListId}")
-    public List<CheckupListDetailModalWheel> checkupListDetailModal(@PathVariable("checkupListId") int checkupListId) {
+    public List<CheckupListDetailModalWheel> checkupListDetailModal(@PathVariable("checkupListId") Long checkupListId) {
 
         return checkupListService.getCheckupListDetailModalWheels(checkupListId);
     }
